`CRTP` completely removed. Even though it's quite useful, if a base class inherits from `CRTP`, then the base class also needs to be templated so it's derived can provide it's typename to process `RTTI`
For example, if `BaseMoveComp` is not templated, and `HeroMoveComp` inherits `BaseMoveComp` `RTTI` will display, `BaseMoveComp` instead of `HeroMoveComp`

To fix this, I had to convert `BaseMoveComp` to fully template however this posing more serious problems (circular dependency) and just a hacky way to fix the root of the problem.

Instead, Factory design pattern employed to fix the problem. After the object is constructed, any arbitrary functions can be automatically called.